modelo-ia/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
onnxruntime>=1.16.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
torch>=2.0.0
torchvision>=0.15.0
